% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model.R
\name{BernHierModel}
\alias{BernHierModel}
\title{Bernoulli hierarchical Beyesian model}
\usage{
BernHierModel(dat, betaA = 2, betaB = 2, numSavedSteps = 1e+05,
  burnInSteps = 4000, adaptSteps = 1000, thinSteps = 1, parallel = TRUE,
  nChains = 4, saveName = "HierModel")
}
\arguments{
\item{dat}{input for the fucntion, should be two colums, first one is read
counts of HapI, second is sum of HapI and HapII}

\item{betaA}{Shape value A for prior beta distribution}

\item{betaB}{Shape value B for prior beta distribution}

\item{numSavedSteps}{total simulation steps}

\item{burnInSteps}{number of burnin steps in MCMC}

\item{adaptSteps}{number of adapting steps}

\item{thinSteps}{thin used in estimating posterior distribution}

\item{parallel}{should MCMC run in parallel?}

\item{nChains}{number of chains used. Maximum is 4}

\item{saveName}{prefix for saved Rdata file}
}
\value{
MCMC simulations results as a coda object
}
\description{
Specify the parameter for hierarchical Beyesian model
}
\note{
Scripts have been modified from "Doing Bayesian Data Analysis (2nd)" by
  John K. Kruschke.
}
\examples{
byesRes.A <- BernHierModel(F1.TypeA[, -1], saveName = "Family1.TypeA")
}

