% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diag.R
\name{diagMCMC}
\alias{diagMCMC}
\title{diagnose plots
Various plots to test the effiency and convergence of MCMC}
\usage{
diagMCMC(codaObject, parName, outputF = "diagPlot")
}
\arguments{
\item{codaObject}{output from \code{\link{BernHierModel}}}

\item{parName}{parameter to be dignosised}

\item{outputF}{output folder}
}
\value{
save diagnosis files in the output folder
}
\description{
diagnose plots
Various plots to test the effiency and convergence of MCMC
}
\note{
Scripts have been modified from "Doing Bayesian Data Analysis (2nd)" by
  John K. Kruschke.
}
\examples{
byesRes.A <- BernHierModel(F1.TypeA[, -1], saveName = "Family1.TypeA")
parameterNames.A <- coda::varnames(byesRes.A)[1]
diagMCMC(byesRes.A, parameterNames.A, outputF = "Family1.TypeA.parameters")
}

