% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NIPT_dat.R
\docType{data}
\name{F1.TypeB}
\alias{F1.TypeB}
\title{Counts data for Type B SNPs}
\format{A data.frame with names of SNPs and counts on HapI and HapII}
\usage{
data(F1.TypeB)
}
\description{
Counts data for Type B SNPs in family one.
}
\examples{
data(F1.TypeB)
}
\references{
Kwan-Wood G. Lam, Peiyong Jiang, et al. Clinical Chemistry (2012)
}
\keyword{datasets}

