% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.R
\name{singlePoster}
\alias{singlePoster}
\title{plot the posterior distribution}
\usage{
singlePoster(codaSamples, para, indiffZone = c(0.49, 0.51), credMass = 0.95,
  showMode = FALSE, usexlim = FALSE, xlimLow = 0.4, xlimHigh = 0.6)
}
\arguments{
\item{codaSamples}{output from \code{\link{BernHierModel}}}

\item{indiffZone}{limits of indiffZone}

\item{credMass}{probility mass for highest density interval (HDI)}

\item{usexlim}{manually set the xlim of histogram?}

\item{xlimLow}{low end of xlim}

\item{xlimHigh}{high end of xlim}
}
\description{
plot the posterior distribution for a single parameter
}
\note{
A wrapper of plotPost fucntion in BEST.
}
\examples{
byesRes.A <- BernHierModel(F1.TypeA[, -1], saveName = "Family1.TypeA")
singlePoster(byesRes.A, "kappa")
}

