% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.R
\name{smryMCMC}
\alias{smryMCMC}
\title{Summarize the posterior distribution}
\usage{
smryMCMC(codaSamples, indiffZone = c(0.49, 0.51), credMass = 0.95,
  usexlim = FALSE, xlimLow = 0.4, xlimHigh = 0.6, saveCSV = TRUE,
  saveName = "HieraraBayesian")
}
\arguments{
\item{codaSamples}{output from \code{\link{BernHierModel}}}

\item{indiffZone}{limits of indiffZone}

\item{credMass}{probility mass for highest density interval (HDI)}

\item{usexlim}{manually set the xlim of histogram?}

\item{xlimLow}{low end of xlim}

\item{xlimHigh}{high end of xlim}

\item{saveCSV}{save the results as a csv file?}

\item{saveName}{prefix for the saved pdf file and csv file}
}
\value{
put the histograms from the posterior distributions into a pdf file
  and return a data.frame contain the summary info
}
\description{
Summarize the posterior distribution for HapBias, kappa and other parameters.
  The posterior distributs will be saved into a pdf file under current folder.
}
\note{
Scripts have been modified from "Doing Bayesian Data Analysis (2nd)" by
  John K. Kruschke.
}
\examples{
byesRes.A <- BernHierModel(F1.TypeA[, -1], saveName = "Family1.TypeA")
summaryRes.A <- smryMCMC(byesRes.A, saveName = "Family1.TypeA")
}

